requirements.txt
tika
pytesseract
pillow
pdf2image